third_party/WebKit/LayoutTests/fast/alignment/parse-align-self-expected.txt

Test that setting and getting align-self works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting align-self set through CSS
PASS getComputedStyle(alignSelfBaseline, '').getPropertyValue('align-self') is 'baseline'
PASS getComputedStyle(alignSelfStretch, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(alignSelfStart, '').getPropertyValue('align-self') is 'start'
PASS getComputedStyle(alignSelfEnd, '').getPropertyValue('align-self') is 'end'
PASS getComputedStyle(alignSelfCenter, '').getPropertyValue('align-self') is 'center'
PASS getComputedStyle(alignSelfSelfEnd, '').getPropertyValue('align-self') is 'self-end'
PASS getComputedStyle(alignSelfSelfStart, '').getPropertyValue('align-self') is 'self-start'
PASS getComputedStyle(alignSelfLeft, '').getPropertyValue('align-self') is 'left'
PASS getComputedStyle(alignSelfRight, '').getPropertyValue('align-self') is 'right'
PASS getComputedStyle(alignSelfEndTrue, '').getPropertyValue('align-self') is 'end true'
PASS getComputedStyle(alignSelfCenterTrue, '').getPropertyValue('align-self') is 'center true'
PASS getComputedStyle(alignSelfSelfEndSafe, '').getPropertyValue('align-self') is 'self-end safe'
PASS getComputedStyle(alignSelfSelfStartSafe, '').getPropertyValue('align-self') is 'self-start safe'
PASS getComputedStyle(alignSelfRightSafe, '').getPropertyValue('align-self') is 'right safe'
PASS getComputedStyle(alignSelfLeftTrue, '').getPropertyValue('align-self') is 'left true'

Test initial value of align-self through JS
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'

Test getting and setting align-self through JS
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'center'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'start true'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'

Test bad combinaisons of align-self
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'

Test the value 'initial'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'center'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'stretch'

Test the value 'inherit'
PASS getComputedStyle(parentElement, '').getPropertyValue('align-self') is 'end'
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'end'
PASS successfullyParsed is true

TEST COMPLETE
